+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 All notable changes to the "persistent-run-buttons" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Unreleased]
+## [1.0.0]
 
 - Initial release
+
+## [1.0.1]
+
+- Updated marketplace display name to "Persistent Run Buttons".
+
+## [1.0.2]
+
+- Added biome config
+- Added compounds to run options
+- Changed to not import the entire vscode namespace
+- Fixed issue with the editor title buttons being split up by other buttons (specifically the open changes button)